# dist/api-docs.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "GC Suite Server",
		"description": "GC Suite Server API docs"
	},
	"basePath": "/api",
	"schemes": [
		"http"
	],
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"paths": {
...
		"/cases": {
			"post": {
				"tags": [
					"Cases"
				],
				"description": "This api creates case",
				"operationId": "/cases",
				"produces": [
					"application/json"
				],
				"parameters": [{
					"name": "Case Info",
					"in": "body",
					"description": "Create case",
					"required": true,
					"schema": {
						"$ref": "#/definitions/createCase"
					}
				}],
				"responses": {
					"200": {}
				}
			},
			"put": {
				"tags": [
					"Cases"
				],
				"description": "This api updates case",
				"operationId": "/cases",
				"produces": [
					"application/json"
				],
				"parameters": [{
					"name": "Case Info",
					"in": "body",
					"description": "Update case",
					"required": true,
					"schema": {
						"$ref": "#/definitions/updateCase"
					}
				}],
				"responses": {
					"200": {}
				}
			},
			"get": {
				"tags": [
					"Cases"
				],
				"description": "This api returns matching cases.",
				"operationId": "/cases",
				"produces": [
					"application/json"
				],
				"parameters": [{
						"name": "_id",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "name",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "tenantId",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "offset",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {}
				}
			},
			"delete": {
				"tags": [
					"Cases"
				],
				"description": "This api delete case.",
				"operationId": "/cases",
				"produces": [
					"application/json"
				],
				"parameters": [{
						"name": "_id",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "tenantId",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {}
				}
			}
		},
...
		"/docs": {
			"post": {
				"tags": [
					"Docs"
				],
				"description": "This api creates docs",
				"operationId": "/docs",
				"produces": [
					"application/json"
				],
				"parameters": [{
					"name": "User Info",
					"in": "body",
					"description": "Create docs",
					"required": true,
					"schema": {
						"$ref": "#/definitions/createDoc"
					}
				}],
				"responses": {
					"200": {}
				}
			},
			"put": {
				"tags": [
					"Docs"
				],
				"description": "This api updates docs",
				"operationId": "/docs",
				"produces": [
					"application/json"
				],
				"parameters": [{
					"name": "User Info",
					"in": "body",
					"description": "Update docs",
					"required": true,
					"schema": {
						"$ref": "#/definitions/updateDoc"
					}
				}],
				"responses": {
					"200": {}
				}
			},
			"get": {
				"tags": [
					"Docs"
				],
				"description": "This api returns matching docs.",
				"operationId": "/docs",
				"produces": [
					"application/json"
				],
				"parameters": [{
						"name": "_id",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "title",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "caseId",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "docId",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "tenantId",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "offset",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {}
				}
			},
			"delete": {
				"tags": [
					"Docs"
				],
				"description": "This api delete doc.",
				"operationId": "/docs",
				"produces": [
					"application/json"
				],
				"parameters": [{
						"name": "_id",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "tenantId",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {}
				}
			}
		},
...
	},
	"definitions": {
...
		"createCase": {
			"properties": {
				"name": {
					"required": true,
					"type": "string"
				},
				"description": {
					"required": false,
					"type": "string"
				},
				"extMatterName": {
					"required": false,
					"type": "string"
				},
				"extMatterNo": {
					"required": false,
					"type": "number"
				},
				"tenantId": {
					"required": true,
					"type": "string"
				}
			}
		},
...
		"createDoc": {
			"properties": {
				"title": {
					"required": false,
					"type": "string"
				},
				"caseId": {
					"required": false,
					"type": "string"
				},
				"docId": {
					"required": false,
					"type": "string"
				},
				"tenantId": {
					"required": true,
					"type": "string"
				}
			}
		},
		"updateDoc": {
			"properties": {
				"_id": {
					"required": true,
					"type": "string"
				},
				"title": {
					"required": false,
					"type": "string"
				},
				"caseId": {
					"required": false,
					"type": "string"
				},
				"docId": {
					"required": false,
					"type": "string"
				},
				"tenantId": {
					"required": true,
					"type": "string"
				}
			}
		},
...
	}
}